refactor: MergePersonaRequest가 String으로 입력받도록 수정한다

+++ src/main/kotlin/org/gitanimals/render/app/UserFacade.kt
@@ -47,8 +47,8 @@ class UserFacade(
 
         return userService.mergePersona(
             user.id.toLong(),
-            request.increasePersonaId,
-            request.deletePersonaId,
+            request.increasePersonaId.toLong(),
+            request.deletePersonaId.toLong(),
         )
     }
 }
